requirements.txt
Django==5.2.5
djangorestframework==3.16.1
asgiref==3.9.1
sqlparse==0.5.3
